# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/": {
      "get": {
        "tags": [
          "Users"
        ],
        "description": "Get all users",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/products/": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Get all products",
        "responses": {
          "200": {
            "description": "Successfully retrieved all products"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Create a new product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createProduct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "description": "Edit an existing product",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editProduct"
            }
          },
          {
            "name": "productId",
            "in": "path",
            "description": "ID number of a specific product",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/products/{productId}": {
      "delete": {
        "tags": [
          "Products"
        ],
        "description": "Soft delete an existing product",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific product"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the product"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Get all categories",
        "responses": {
          "200": {
            "description": "Successfully retrieved all categories"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Create a new category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createCategory"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Categories"
        ],
        "description": "Edit an existing category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editCategory"
            }
          },
          {
            "name": "CategoryId",
            "in": "path",
            "description": "ID number of a specific category",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/categories/{CategoryId}": {
      "delete": {
        "tags": [
          "Categories"
        ],
        "description": "Delete an existing category",
        "parameters": [
          {
            "name": "CategoryId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific category"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the category"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/brands/": {
      "get": {
        "tags": [
          "Brands"
        ],
        "description": "Get all brands",
        "responses": {
          "200": {
            "description": "Successfully retrieved all brands"
          },
          "400": {
            "description": "Fail! Something went wrong"
          }
        }
      },
      "post": {
        "tags": [
          "Brands"
        ],
        "description": "Create a new brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Brands"
        ],
        "description": "Edit an existing brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editBrand"
            }
          },
          {
            "name": "BrandId",
            "in": "path",
            "description": "ID number of a specific brand",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/brands/{BrandId}": {
      "delete": {
        "tags": [
          "Brands"
        ],
        "description": "",
        "parameters": [
          {
            "name": "BrandId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "ID number of a specific brand"
          },
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/orders/": {
      "get": {
        "tags": [
          "Orders"
        ],
        "description": "Get all orders",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all orders"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "description": "Edit an existing order status",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/editOrder"
            }
          },
          {
            "name": "OrderId",
            "in": "path",
            "description": "ID number of a specific order",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully edited the orders status"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "Get admin dashboard data",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved admin dashboard data"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/admin/brands": {
      "get": {
        "tags": [
          "Admin-Brands"
        ],
        "description": "Get all brands",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all brands"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "tags": [
          "Admin-Brands"
        ],
        "description": "Create a new brand",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/createBrand"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully created a new brand"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
    "/admin/orders": {
      "get": {
        "tags": [
          "Admin-Orders"
        ],
        "description": "Get all orders",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "Admin Only! Bearer token to authorize the request",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all orders"
          },
          "400": {
            "description": "Fail! Something went wrong"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createBrand": {
      "type": "object",
      "properties": {
        "Brand": {
          "type": "string",
          "example": "Example Brand"
        }
      },
      "required": [
        "Brand"
      ]
    },
...
    "createCategory": {
      "type": "object",
      "properties": {
        "Category": {
          "type": "string",
          "example": "Example Category"
        }
      },
      "required": [
        "Category"
      ]
    },
    "editCategory": {
      "type": "object",
      "properties": {
        "NewCategory": {
          "type": "string",
          "example": "Edited Category"
        }
      },
      "required": [
        "NewCategory"
      ]
    },
    "createProduct": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Product Example"
        },
        "Description": {
          "type": "string",
          "example": "Product description ...."
        },
        "UnitPrice": {
          "type": "number",
          "example": 15
        },
        "ImageURL": {
          "type": "string",
          "example": "http://images.restapi.co.za/products/product-apple-tv.png"
        },
        "Quantity": {
          "type": "number",
          "example": 5
        },
        "BrandName": {
          "type": "string",
          "example": "Example Brand"
        },
        "CategoryName": {
          "type": "string",
          "example": "Example Category"
        }
      },
      "required": [
        "Name",
        "Description",
        "UnitPrice",
        "ImageURL",
        "Quantity",
        "BrandName",
        "CategoryName"
      ]
    },
...
  }
}